requirements.txt
flask>=3.0.0
tensorflow>=2.11.0
numpy>=1.23.0
pillow>=10.0.0
matplotlib>=3.7.0
opencv-python>=4.8.0
gunicorn>=21.2.0
